+++ examples/more/header-only-defines/application/main.cpp
@@ -1,0 +1,9 @@
+#include <cstdio>
+
+#include "favorite-color/color.h"
+
+int main(int, char**)
+{
+    printf("%s\n", color::favorite());
+    return 0;
+}